fix(react) - Pass in component directly as the tab content
