+++ packages/react/src/components/Tabs/Tabs.props.ts
@@ -1,5 +1,4 @@
 import { IconProps } from "../Icon/Icon.props";
-
 export interface TabsProps {
   /**
    * Specify the items inside a tab
@@ -10,6 +9,5 @@ export interface TabsProps {
 export interface TabItem {
   icon?: IconProps;
   label: Required<string>;
-  component: Required<string>;
-  componentData: Required<any>;
+  content: Required<any>;
 }
